# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Infinity Data Explorer API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/factions/{slug}": {
      "get": {
        "tags": [
          "factions"
        ],
        "summary": "Get Faction",
        "description": "Get faction details including unit count and fireteam chart.",
        "operationId": "get_faction_api_factions__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FactionDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/factions/{slug}/legacy": {
      "get": {
        "tags": [
          "factions"
        ],
        "summary": "Get Faction Legacy",
        "description": "Return raw JSON blob exactly matching frontend legacy layout.",
        "operationId": "get_faction_legacy_api_factions__slug__legacy_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/lists": {
      "get": {
        "tags": [
          "Lists"
        ],
        "summary": "Get User Lists",
        "description": "Get all saved army lists for the authenticated user.",
        "operationId": "get_user_lists_api_lists_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ArmyListSummaryResponse"
                  },
                  "type": "array",
                  "title": "Response Get User Lists Api Lists Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Lists"
        ],
        "summary": "Create List",
        "description": "Create a new army list for the user.",
        "operationId": "create_list_api_lists_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArmyListCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArmyListDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/lists/{list_id}": {
      "get": {
        "tags": [
          "Lists"
        ],
        "summary": "Get List",
        "description": "Get details of a specific army list.",
        "operationId": "get_list_api_lists__list_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "list_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "List Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArmyListDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Lists"
        ],
        "summary": "Update List",
        "description": "Update an existing army list.",
        "operationId": "update_list_api_lists__list_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "list_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "List Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArmyListUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArmyListDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Lists"
        ],
        "summary": "Delete List",
        "description": "Delete an army list.",
        "operationId": "delete_list_api_lists__list_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "list_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "List Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArmyListCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "faction_id": {
            "type": "integer",
            "title": "Faction Id"
          },
          "points": {
            "type": "integer",
            "title": "Points",
            "default": 0
          },
          "swc": {
            "type": "number",
            "title": "Swc",
            "default": 0.0
          },
          "units_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Units Json",
            "description": "JSON payload representing the list composition"
          }
        },
        "type": "object",
        "required": [
          "name",
          "faction_id"
        ],
        "title": "ArmyListCreate"
      },
      "ArmyListDetailResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "faction_id": {
            "type": "integer",
            "title": "Faction Id"
          },
          "points": {
            "type": "integer",
            "title": "Points"
          },
          "swc": {
            "type": "number",
            "title": "Swc"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "units_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Units Json"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "faction_id",
          "points",
          "swc",
          "created_at",
          "updated_at",
          "units_json"
        ],
        "title": "ArmyListDetailResponse"
      },
      "ArmyListSummaryResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "faction_id": {
            "type": "integer",
            "title": "Faction Id"
          },
          "points": {
            "type": "integer",
            "title": "Points"
          },
          "swc": {
            "type": "number",
            "title": "Swc"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "faction_id",
          "points",
          "swc",
          "created_at",
          "updated_at"
        ],
        "title": "ArmyListSummaryResponse"
      },
      "ArmyListUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "faction_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Faction Id"
          },
          "points": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Points"
          },
          "swc": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Swc"
          },
          "units_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Units Json"
          }
        },
        "type": "object",
        "title": "ArmyListUpdate"
      },
...
      "FactionDetailResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "is_vanilla": {
            "type": "boolean",
            "title": "Is Vanilla"
          },
          "discontinued": {
            "type": "boolean",
            "title": "Discontinued"
          },
          "logo": {
            "type": "string",
            "title": "Logo"
          },
          "fireteam_chart": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fireteam Chart"
          },
          "unit_count": {
            "type": "integer",
            "title": "Unit Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "slug",
          "parent_id",
          "is_vanilla",
          "discontinued",
          "logo"
        ],
        "title": "FactionDetailResponse",
        "description": "Full faction detail including fireteam chart."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}